refactor(gateway): Separate business logic from the controller to a new service.
Added:
- src/modules/transactions/service.js

Updated:
- src/modules/transactions/controller.js
