+++ gateway/src/modules/transactions/controller.js
@@ -1,15 +1,12 @@
-const { TransactionRequestPublisher } = require("@yape-challenge/kafka");
-const { v4: uuidv4 } = require("uuid");
+const { TransactionService } = require('./service');
 
 async function createTransaction(req, res, next) {
   try {
     const { body } = req;
-    const correlationId = uuidv4();
-    const transaction = { ...body, correlationId };
 
-    await TransactionRequestPublisher.publish(transaction);
+    const transaction = await TransactionService.publishTransaction(body);
 
-    res.status(201).json({ message: "Transaction created", data: transaction });
+    res.status(201).json({ message: 'Transaction created', data: transaction });
   } catch (error) {
     next(error);
   }
@@ -18,17 +15,11 @@ async function createTransaction(req, res, next) {
 async function createMassiveTransactions(req, res, next) {
   try {
     const { body } = req;
-    const groupId = uuidv4();
 
-    const publishPromises = body.map((transaction) => {
-      const message = { ...transaction, correlationId: groupId };
-      return TransactionRequestPublisher.publish(message);
-    });
-
-    await Promise.all(publishPromises);
+    const groupId = await TransactionService.publishMasiveTransaction(body);
 
     res.status(201).json({
-      message: "Massive transactions created",
+      message: 'Massive transactions created',
       data: {
         correlationId: groupId,
       },
